refactor: Reorg the repo
