+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.25.1] - 05-07-23
+### Changed
+
+* Refactored package into `bootstrap`, `clients` and `preprocessor` modules. This is likely a breaking change for most users of the package.
+
+
 ## [0.24.1] - 30-06-23
 ### Removed
