+++ cognite/powerops/_shared_data_classes.py
