+++ cognite/powerops/clients/shop/__init__.py
@@ -1,0 +1,3 @@
+from ._api_client import ShopClient
+
+__all__ = ["ShopClient"]
