+++ cognite/powerops/clients/shop/data_classes/__init__.py
@@ -1,8 +1,8 @@
-from cognite.powerops.client.data_classes.case import Case
-from cognite.powerops.client.data_classes.shop_result_files import ShopLogFile, ShopResultFile, ShopYamlFile
-from cognite.powerops.client.data_classes.shop_results import ObjectiveFunction, ShopRunResult
-from cognite.powerops.client.data_classes.shop_run import ShopRun
-from cognite.powerops.client.data_classes.shop_run_event import ShopRunEvent
+from cognite.powerops.clients.shop.data_classes.case import Case
+from cognite.powerops.clients.shop.data_classes.shop_result_files import ShopLogFile, ShopResultFile, ShopYamlFile
+from cognite.powerops.clients.shop.data_classes.shop_results import ObjectiveFunction, ShopRunResult
+from cognite.powerops.clients.shop.data_classes.shop_run import ShopRun
+from cognite.powerops.clients.shop.data_classes.shop_run_event import ShopRunEvent
 
 __all__ = [
     "Case",
