+++ cognite/powerops/clients/shop/data_classes/shop_results_compare.py
@@ -11,12 +11,12 @@ from deepdiff import DeepDiff
 from deepdiff.model import PrettyOrderedSet
 
 from cognite.powerops.bootstrap.utils.plotting import ax_plot_time_series, create_time_series_plot
-from cognite.powerops.client.data_classes.helpers import (
+from cognite.powerops.clients.shop.data_classes.helpers import (
     format_deep_diff_path,
     get_data_from_nested_dict,
     is_time_series_dict,
 )
-from cognite.powerops.client.data_classes.shop_result_files import ShopYamlFile
+from cognite.powerops.clients.shop.data_classes.shop_result_files import ShopYamlFile
 
 logger = logging.getLogger(__name__)
