+++ cognite/powerops/clients/shop/data_classes/shop_run.py
@@ -5,8 +5,8 @@ import time
 from enum import Enum
 from typing import TYPE_CHECKING
 
-from cognite.powerops.client.data_classes.shop_results import ShopRunResult
-from cognite.powerops.client.data_classes.shop_run_event import ShopRunEvent
+from cognite.powerops.clients.shop.data_classes.shop_results import ShopRunResult
+from cognite.powerops.clients.shop.data_classes.shop_run_event import ShopRunEvent
 
 if TYPE_CHECKING:
     from cognite.powerops import PowerOpsClient
