+++ cognite/powerops/utils/cdf/calls.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import logging
-from typing import Optional, Union, cast
+from typing import Optional, Union
 
 from cognite.client import CogniteClient
 from cognite.client._api.assets import AssetsAPI
@@ -10,54 +10,10 @@ from cognite.client._api.files import FilesAPI
 from cognite.client._api.labels import LabelsAPI
 from cognite.client._api.relationships import RelationshipsAPI
 from cognite.client._api.sequences import SequencesAPI
-from cognite.client.data_classes import (
-    Asset,
-    DataSet,
-    Event,
-    FileMetadata,
-    LabelDefinition,
-    Relationship,
-    RelationshipList,
-    Sequence,
-    TimeSeries,
-)
+from cognite.client.data_classes import Asset, Event, LabelDefinition, Relationship, RelationshipList, Sequence
 from cognite.client.exceptions import CogniteDuplicatedError
 
-from cognite.powerops.client.dm_client._api._core import TypeAPI
-from cognite.powerops.client.dm_client.data_classes import (
-    FileRef,
-    FileRefApply,
-    Mapping,
-    MappingApply,
-    ModelTemplate,
-    ModelTemplateApply,
-    Transformation,
-    TransformationApply,
-)
-from cognite.powerops.client.dm_client.data_classes._core import DomainModel, DomainModelApply
-
 logger = logging.getLogger(__name__)
-
-
-def simple_relationship(
-    source: Union[Asset, TimeSeries, FileMetadata, Sequence, Event],
-    target: Union[Asset, TimeSeries, FileMetadata, Sequence, Event],
-    label_external_id: str,
-) -> Relationship:
-    """Simplifies Cognite Python SDK creation of Relationships."""
-    external_id = f"{source.external_id}.{target.external_id}"
-    source_type = "file" if isinstance(source, FileMetadata) else source.__class__.__name__
-    target_type = "file" if isinstance(target, FileMetadata) else target.__class__.__name__
-
-    return Relationship(
-        external_id=external_id,
-        source_type=source_type,
-        target_type=target_type,
-        source_external_id=source.external_id,
-        target_external_id=target.external_id,
-        labels=[label_external_id],
-    )
-
 
 CogniteAPI = Union[
     AssetsAPI,
@@ -95,21 +51,6 @@ def upsert_cognite_resources(
             api.create(resources_to_update)
 
 
-def upsert_cognite_dm_resources(
-    api: TypeAPI,
-    resources: list[DomainModelApply],
-) -> None:
-    for resource in resources:
-        api.apply(resource, replace=True)
-
-
-def retrieve_dataset(client: CogniteClient, external_id: str) -> DataSet:
-    dataset = client.data_sets.retrieve(external_id=external_id)
-    if dataset is None:
-        raise ValueError(f"DataSet not found: {external_id}")
-    return dataset
-
-
 def retrieve_event(client: CogniteClient, external_id: str) -> Event:
     event = client.events.retrieve(external_id=external_id)
     if event is None:
@@ -138,25 +79,3 @@ def retrieve_relationships_from_source_ext_id(
         limit=-1,
         target_types=target_types,
     )
-
-
-def to_dm_apply(instances: list[DomainModel] | DomainModel) -> list[DomainModelApply]:
-    items = cast(list[DomainModel], list(instances))
-    apply_items = []
-
-    type_map = {
-        ModelTemplate: ModelTemplateApply,
-        Mapping: MappingApply,
-        Transformation: TransformationApply,
-        FileRef: FileRefApply,
-    }
-
-    for item in items:
-        if type(item) not in type_map:
-            raise NotImplementedError(f"Dont know how to convert {item!r}.")
-        apply_type = type_map[type(item)]
-
-        apply_items.append(
-            apply_type(**{field: value for field, value in item.dict().items() if field in apply_type.__fields__})
-        )
-    return apply_items
