+++ cognite/powerops/utils/retry/__init__.py
@@ -1,5 +1,3 @@
-__all__ = ["retry", "retry_call"]
-
 import logging
 
 from .api import retry, retry_call
@@ -7,3 +5,5 @@ from .api import retry, retry_call
 # Set default logging handler to avoid "No handler found" warnings.
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
+
+__all__ = ["retry", "retry_call"]
