+++ tests/test_unit/test_case/test_case_loading.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 import yaml
 
-from cognite.powerops.client.data_classes import Case
+from cognite.powerops.clients.shop.data_classes import Case
 
 
 @pytest.fixture
